+++ IsekaiMod/Content/Classes/IsekaiProtagonist/Archetypes/GodEmporer.cs
@@ -12,6 +12,7 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist.Archetypes
             var GodEmperorProficiencies = Resources.GetModBlueprint<BlueprintFeature>("GodEmperorProficiencies");
             var NascentApotheosis = Resources.GetModBlueprint<BlueprintFeature>("NascentApotheosis");
             var DivineArray = Resources.GetModBlueprint<BlueprintFeature>("DivineArray");
+            var GodEmperorEnergySelection = Resources.GetModBlueprint<BlueprintFeature>("GodEmperorEnergySelection");
             var ProtectiveAuraFeature = Resources.GetModBlueprint<BlueprintFeature>("ProtectiveAuraFeature");
             var DarkAuraFeature = Resources.GetModBlueprint<BlueprintFeature>("DarkAuraFeature");
             var GloriousAuraFeature = Resources.GetModBlueprint<BlueprintFeature>("GloriousAuraFeature");
@@ -57,6 +58,7 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist.Archetypes
                 bp.AddFeatures = new LevelEntry[] {
                     Helpers.LevelEntry(1, NascentApotheosis, GodEmperorProficiencies),
                     Helpers.LevelEntry(3, DivineArray),
+                    Helpers.LevelEntry(5, GodEmperorEnergySelection),
                     Helpers.LevelEntry(7, ProtectiveAuraFeature),
                     Helpers.LevelEntry(9, GloriousAuraFeature),
                     Helpers.LevelEntry(10, DarkAuraFeature),