+++ IsekaiMod/Content/ContentAdder.cs
@@ -96,6 +96,7 @@ namespace IsekaiMod.Content
                 Features.IsekaiProtagonist.Archetypes.GodEmperor.GodEmperorProficiencies.Add();
                 Features.IsekaiProtagonist.Archetypes.GodEmperor.NascentApotheosis.Add();
                 Features.IsekaiProtagonist.Archetypes.GodEmperor.DivineArray.Add();
+                Features.IsekaiProtagonist.Archetypes.GodEmperor.GodEmperorEnergySelection.Add();
                 Features.IsekaiProtagonist.Archetypes.GodEmperor.ProtectiveAuraFeature.Add();
                 Features.IsekaiProtagonist.Archetypes.GodEmperor.DarkAuraFeature.Add();
                 Features.IsekaiProtagonist.Archetypes.GodEmperor.GloriousAuraFeature.Add();